+++ backend/src/main/java/com/brokendev/backend/account/dto/AccountDepositResponseDTO.java
@@ -1,4 +1,4 @@
-package com.brokendev.backend.dto.account;
+package com.brokendev.backend.account.dto;
 
 import java.math.BigDecimal;
