+++ backend/src/main/java/com/brokendev/backend/account/dto/TransactionStatementResponseDTO.java
@@ -1,4 +1,4 @@
-package com.brokendev.backend.dto.account;
+package com.brokendev.backend.account.dto;
 
 import com.brokendev.backend.enums.TransactionType;
