+++ backend/src/main/java/com/brokendev/backend/auth/service/AuthService.java
@@ -4,12 +4,12 @@ import com.brokendev.backend.auth.dto.login.LoginRequestDTO;
 import com.brokendev.backend.auth.dto.login.LoginResponseDTO;
 import com.brokendev.backend.auth.dto.register.RegisterRequestDTO;
 import com.brokendev.backend.auth.dto.register.RegisterResponseDTO;
-import com.brokendev.backend.common.domain.account.Account;
+import com.brokendev.backend.account.domain.Account;
 import com.brokendev.backend.common.domain.user.User;
 import com.brokendev.backend.common.exceptions.InvalidPasswordException;
 import com.brokendev.backend.common.exceptions.UserAlreadyExistsException;
 import com.brokendev.backend.infra.security.TokenService;
-import com.brokendev.backend.common.domain.account.AccountRepository;
+import com.brokendev.backend.account.domain.AccountRepository;
 import com.brokendev.backend.common.domain.user.UserRepository;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
