+++ backend/src/main/java/com/brokendev/backend/boleto_payment/domain/BoletoPayment.java
@@ -1,6 +1,6 @@
-package com.brokendev.backend.domain;
+package com.brokendev.backend.boleto_payment.domain;
 
-import com.brokendev.backend.common.domain.account.Account;
+import com.brokendev.backend.account.domain.Account;
 import com.brokendev.backend.enums.BoletoPaymentStatus;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
