+++ backend/src/main/java/com/brokendev/backend/dto/dashboard/DashboardResponseDTO.java
@@ -1,6 +1,6 @@
 package com.brokendev.backend.dto.dashboard;
 
-import com.brokendev.backend.dto.account.TransactionStatementResponseDTO;
+import com.brokendev.backend.account.dto.TransactionStatementResponseDTO;
 
 import java.math.BigDecimal;
 import java.util.List;
