+++ backend/src/main/java/com/brokendev/backend/dto/profile/UserProfileResponseDTO.java
@@ -1,6 +1,6 @@
 package com.brokendev.backend.dto.profile;
 
-import com.brokendev.backend.dto.account.AccountInfoResponseDTO;
+import com.brokendev.backend.account.dto.AccountInfoResponseDTO;
 import com.brokendev.backend.dto.card.CardResponseDTO;
 
 import java.util.List;
