+++ backend/src/main/java/com/brokendev/backend/pix_transfer/domain/PixTransaction.java
@@ -1,7 +1,7 @@
-package com.brokendev.backend.domain;
+package com.brokendev.backend.pix_transfer.domain;
 
 
-import com.brokendev.backend.common.domain.account.Account;
+import com.brokendev.backend.account.domain.Account;
 import com.brokendev.backend.enums.PixKeyType;
 import com.brokendev.backend.enums.PixTransactionStatus;
 import jakarta.persistence.*;
