+++ backend/src/main/java/com/brokendev/backend/services/DashboardService.java
@@ -1,9 +1,9 @@
 package com.brokendev.backend.services;
 
-import com.brokendev.backend.dto.account.TransactionStatementResponseDTO;
+import com.brokendev.backend.account.dto.TransactionStatementResponseDTO;
 import com.brokendev.backend.dto.dashboard.DashboardResponseDTO;
 import com.brokendev.backend.common.exceptions.AccountNotFoundException;
-import com.brokendev.backend.common.domain.account.AccountRepository;
+import com.brokendev.backend.account.domain.AccountRepository;
 import com.brokendev.backend.common.domain.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
