+++ backend/src/main/java/com/brokendev/backend/services/UserService.java
@@ -1,15 +1,15 @@
 package com.brokendev.backend.services;
 
-import com.brokendev.backend.common.domain.account.Account;
+import com.brokendev.backend.account.domain.Account;
 import com.brokendev.backend.common.domain.user.User;
-import com.brokendev.backend.dto.account.AccountInfoResponseDTO;
+import com.brokendev.backend.account.dto.AccountInfoResponseDTO;
 import com.brokendev.backend.dto.card.CardResponseDTO;
 import com.brokendev.backend.dto.profile.UserProfileResponseDTO;
 import com.brokendev.backend.dto.profile.UserProfileUpdateDTO;
 import com.brokendev.backend.dto.profile.UserProfileUpdateResponseDTO;
 import com.brokendev.backend.common.exceptions.AccountNotFoundException;
 import com.brokendev.backend.infra.security.TokenService;
-import com.brokendev.backend.common.domain.account.AccountRepository;
+import com.brokendev.backend.account.domain.AccountRepository;
 import com.brokendev.backend.repositories.CardRepository;
 import com.brokendev.backend.common.domain.user.UserRepository;
 import jakarta.transaction.Transactional;
