+++ backend/src/test/java/com/brokendev/backend/services/AccountServiceTest.java
@@ -1,20 +1,22 @@
 package com.brokendev.backend.services;
 
-import com.brokendev.backend.common.domain.account.Account;
-import com.brokendev.backend.common.domain.account.AccountRepository;
+import com.brokendev.backend.account.domain.Account;
+import com.brokendev.backend.account.domain.AccountRepository;
+import com.brokendev.backend.boleto_payment.domain.BoletoPaymentRepository;
 import com.brokendev.backend.common.domain.user.UserRepository;
-import com.brokendev.backend.domain.BoletoPayment;
+import com.brokendev.backend.boleto_payment.domain.BoletoPayment;
 import com.brokendev.backend.common.domain.user.User;
-import com.brokendev.backend.dto.account.AccountBalanceResponseDTO;
-import com.brokendev.backend.dto.account.AccountDepositResponseDTO;
-import com.brokendev.backend.dto.boleto.BoletoPaymentRequestDTO;
-import com.brokendev.backend.dto.boleto.BoletoPaymentResponseDTO;
-import com.brokendev.backend.dto.pixTransfer.PixTransferRequestDTO;
+import com.brokendev.backend.account.dto.AccountBalanceResponseDTO;
+import com.brokendev.backend.account.dto.AccountDepositResponseDTO;
+import com.brokendev.backend.boleto_payment.dto.BoletoPaymentRequestDTO;
+import com.brokendev.backend.boleto_payment.dto.BoletoPaymentResponseDTO;
+import com.brokendev.backend.pix_transfer.dto.PixTransferRequestDTO;
 import com.brokendev.backend.enums.BoletoPaymentStatus;
 import com.brokendev.backend.enums.PixKeyType;
 import com.brokendev.backend.common.exceptions.AccountNotFoundException;
 import com.brokendev.backend.common.exceptions.InsufficientBalanceException;
 import com.brokendev.backend.common.exceptions.PixTransferNotAllowedException;
+import com.brokendev.backend.pix_transfer.domain.PixTransactionRepository;
 import com.brokendev.backend.repositories.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
