+++ backend/src/test/java/com/brokendev/backend/services/DashboardServiceTest.java
@@ -1,11 +1,11 @@
 package com.brokendev.backend.services;
 
 import com.brokendev.backend.common.domain.user.User;
-import com.brokendev.backend.common.domain.account.Account;
-import com.brokendev.backend.dto.account.TransactionStatementResponseDTO;
+import com.brokendev.backend.account.domain.Account;
+import com.brokendev.backend.account.dto.TransactionStatementResponseDTO;
 import com.brokendev.backend.dto.dashboard.DashboardResponseDTO;
 import com.brokendev.backend.common.exceptions.AccountNotFoundException;
-import com.brokendev.backend.common.domain.account.AccountRepository;
+import com.brokendev.backend.account.domain.AccountRepository;
 import com.brokendev.backend.common.domain.user.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
